Moved semicolon for formality of programming
A semicolon belongs after a variable assignment, but it does not belong after a function definition.  Although no errors were reported and the program was running just fine, I just happened to notice the fact that this semicolon had been misplaced.  It was in line 42, at the end of a function definition, whereas there was no semicolon in line 46, at the end of a statement in which a function was assigned to a variable.